--- client/src/Components/Navigation.js
@@ -1,30 +1,0 @@
-import React from 'react';
-import Container from 'react-bootstrap/Container';
-import Nav from 'react-bootstrap/Nav';
-import Navbar from 'react-bootstrap/Navbar';
-
-
-const styles={
-  h1:{
-    color:"red",
-    }
-}
-
-function Navigation() {
-  return (
-    <div>
-  <Navbar bg="dark" variant="dark">
-        <Container>
-          <Navbar.Brand href="#home">Navbar</Navbar.Brand>
-          <Nav className="me-auto">
-            <Nav.Link href="#home">Home</Nav.Link>
-            <Nav.Link href="#features">Features</Nav.Link>
-            <Nav.Link href="#pricing">Pricing</Nav.Link>
-          </Nav>
-        </Container>
-</Navbar>
-    </div>
-  );
-}
-
-export default Navigation;
